refactor(typehints-gp[resolvers]) Factor shared catalog into _resolvers
why: After D3 and D4 each landed with their own copies of
ResolveContext, Resolver Protocol, and run_chain, the duplication
was ready to be factored into one home. The empirical inventory
in notes/defaults-discovery-d1.md shows the workspace's resolver
needs are homogeneous (factory + truncation), so this is also the
right moment to decide on a public API: hold private behind the
underscore-prefixed module name pending external demand. Migrating
to a public surface later is mechanical (one app.add_default_resolver
call exposed via extension.setup) and the evidence doesn't justify
shipping that maintenance surface speculatively.

what:
- Add `_resolvers.py` hosting `ResolveContext`, `Resolver`,
  `run_chain`, `DataclassFactoryRepr`, and `TruncateLongRepr`.
- Slim `_param_defaults.py` down to its listener
  (`update_synthetic_defvalues`, `_walk_to_dataclass`) and its
  resolver tuple (`_DEFAULT_RESOLVERS = (DataclassFactoryRepr(),)`)
  — imports the shared classes.
- Slim `_data_defaults.py` similarly: keeps
  `Gp{Data,Attribute}Documenter`, `_curate_value_line`, and its
  resolver tuple; imports `TruncateLongRepr` from the shared
  module.
- Update tests to import from `_resolvers` so the assertions
  exercise the canonical home.
- Add `notes/defaults-discovery-d5.md` documenting the factoring
  outcome plus the "hold public API private" decision and the
  triggers that should make us revisit it.
- Behavior unchanged. Same 1 536 tests pass; same docs build.
