+++ packages/sphinx-autodoc-typehints-gp/src/sphinx_autodoc_typehints_gp/_param_defaults.py
@@ -35,134 +35,19 @@ import typing as t
 
 from sphinx.util.inspect import DefaultValue
 
+from sphinx_autodoc_typehints_gp._resolvers import (
+    DataclassFactoryRepr,
+    ResolveContext,
+    Resolver,
+    run_chain,
+)
+
 if t.TYPE_CHECKING:
     from sphinx.application import Sphinx
 
 logger = logging.getLogger(__name__)
 
-
-class ResolveContext(t.NamedTuple):
-    """Context passed to each :class:`Resolver` in the chain.
-
-    Parameters
-    ----------
-    value : object
-        The live Python object (a callable for ``default_factory``
-        cases, or the raw default for direct-value cases).
-    kind : str
-        One of ``'param'``, ``'data'``, ``'attribute'``. Stage B
-        listeners only emit ``'param'``; Stage A's
-        ``GpDataDocumenter`` (D4) emits ``'data'`` / ``'attribute'``.
-    qualname : str
-        Fully qualified name of the documented object, e.g.
-        ``'libtmux.constants.HookEventDataclass.__init__'``.
-    param_name : str | None
-        Set when ``kind == 'param'``; the parameter name being
-        resolved.
-    default_repr : str
-        Sphinx's stock ``object_description`` of ``value``, available
-        as a fallback for resolvers that want to compose with the
-        default rendering.
-    """
-
-    value: object
-    kind: str
-    qualname: str
-    param_name: str | None
-    default_repr: str
-
-
-class Resolver(t.Protocol):
-    """Compute a symbolic source-text string for a default value.
-
-    Resolvers are run in priority order; the first non-``None``
-    result wins. Return ``None`` to defer to the next resolver.
-    Return an empty string to suppress the default (Site A only;
-    parameter defaults cannot be suppressed).
-    """
-
-    def __call__(self, ctx: ResolveContext) -> str | None:
-        """Return the chosen text or ``None`` to defer."""
-        ...
-
-
-class DataclassFactoryRepr:
-    """Render :func:`dataclasses.field` ``default_factory`` symbolically.
-
-    Recognises stdlib container constructors (``list``, ``dict``,
-    ``set``, ``frozenset``, ``tuple``) and named callable types.
-    Defers (returns ``None``) on lambdas and unrecognised factories,
-    leaving Sphinx's stock ``<factory>`` rendering in place.
-
-    Examples
-    --------
-    >>> r = DataclassFactoryRepr()
-    >>> ctx = ResolveContext(
-    ...     value=list,
-    ...     kind='param',
-    ...     qualname='Foo.__init__',
-    ...     param_name='items',
-    ...     default_repr='<factory>',
-    ... )
-    >>> r(ctx)
-    '[]'
-    >>> r(ctx._replace(value=dict))
-    '{}'
-    >>> r(ctx._replace(value=set))
-    'set()'
-    >>> r(ctx._replace(value=lambda: 1)) is None
-    True
-    """
-
-    _BUILTIN_LITERALS: t.ClassVar[dict[type, str]] = {
-        list: "[]",
-        dict: "{}",
-        set: "set()",
-        frozenset: "frozenset()",
-        tuple: "()",
-    }
-
-    def __call__(self, ctx: ResolveContext) -> str | None:
-        """Resolve a ``default_factory`` callable to its source text."""
-        if ctx.kind != "param":
-            return None
-        factory = ctx.value
-        if isinstance(factory, type):
-            literal = self._BUILTIN_LITERALS.get(factory)
-            if literal is not None:
-                return literal
-            name = getattr(factory, "__name__", None)
-            if name and name != "<lambda>":
-                return f"{name}()"
-        return None
-
-
 _DEFAULT_RESOLVERS: tuple[Resolver, ...] = (DataclassFactoryRepr(),)
-
-
-def _run_chain(
-    ctx: ResolveContext,
-    resolvers: tuple[Resolver, ...],
-) -> str | None:
-    """Run *resolvers* in order and return the first non-``None`` result.
-
-    Examples
-    --------
-    >>> ctx = ResolveContext(
-    ...     value=list,
-    ...     kind='param',
-    ...     qualname='Foo.__init__',
-    ...     param_name='items',
-    ...     default_repr='<factory>',
-    ... )
-    >>> _run_chain(ctx, _DEFAULT_RESOLVERS)
-    '[]'
-    """
-    for resolver in resolvers:
-        result = resolver(ctx)
-        if result is not None:
-            return result
-    return None
 
 
 def _walk_to_dataclass(obj: t.Any) -> type | None:
@@ -275,7 +160,7 @@ def update_synthetic_defvalues(
             param_name=param.name,
             default_repr="<factory>",
         )
-        text = _run_chain(ctx, _DEFAULT_RESOLVERS)
+        text = run_chain(ctx, _DEFAULT_RESOLVERS)
         if text is None:
             new_parameters.append(param)
             continue
